requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.3.0
kaggle>=1.5.0
